myproject/requirements.txt
django>=4.2
djangorestframework>=3.15
openai>=1.40
pydantic>=2.6
python-dotenv>=1.0